# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NetVent Parivaar API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/membership/apply:
    post:
      operationId: applyMembership
      tags: [membership]
      summary: Submit membership application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/membership/me:
    patch:
      operationId: updateMyProfile
      tags: [membership]
      summary: Update own profile and privacy settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberProfile"

  # EXPERTISE LOOKUP
...
  /v1/users/me:
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email or phone already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/users/me/availability:
    patch:
      operationId: setAvailability
      tags: [auth]
      summary: Update own availability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # DIRECTORY
  /v1/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: Search and list members
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: expertise
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectoryListResponse"
  /v1/members/{memberId}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get member profile
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectoryMemberResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # EVENTS
...
  /v1/reports:
    post:
      operationId: createReport
      tags: [members]
      summary: Report a post or profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # AUTH SCHEMAS
...
    UserMe:
      type: object
      required: [id, name, email, role, isAvailable]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [BASIC, MEMBER, MODERATOR, ADMIN]
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        educationType:
          type: ["string", "null"]
          enum: [SCHOOL, COLLEGE, null]
        schoolOrCollegeName:
          type: ["string", "null"]
        schoolClass:
          type: ["string", "null"]
        degreeLevel:
          type: ["string", "null"]
          enum: [DIPLOMA, BACHELORS, MASTERS, null]
        collegeYear:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        expertise:
          type: array
          items:
            $ref: "#/components/schemas/Expertise"
        memberProfile:
          $ref: "#/components/schemas/MemberProfile"
    AuthResponse:
      type: object
      required: [success, user]
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/UserMe"
    Expertise:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
...
    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: "^[0-9]{10}$"
        city:
          type: string
        educationType:
          type: string
          enum: [SCHOOL, COLLEGE]
        schoolOrCollegeName:
          type: ["string", "null"]
        schoolClass:
          type: ["string", "null"]
        degreeLevel:
          type: ["string", "null"]
          enum: [DIPLOMA, BACHELORS, MASTERS, null]
        collegeYear:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        password:
          type: string
          minLength: 8
        expertiseIds:
          type: array
          items:
            type: string
          maxItems: 3
    AvailabilityInput:
      type: object
      required: [isAvailable]
      properties:
        isAvailable:
          type: boolean
    DirectoryMember:
      type: object
      required: [id, name, isAvailable, expertise]
      properties:
        id:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        educationType:
          type: ["string", "null"]
          enum: [SCHOOL, COLLEGE, null]
        schoolOrCollegeName:
          type: ["string", "null"]
        schoolClass:
          type: ["string", "null"]
        degreeLevel:
          type: ["string", "null"]
          enum: [DIPLOMA, BACHELORS, MASTERS, null]
        collegeYear:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        expertise:
          type: array
          items:
            $ref: "#/components/schemas/Expertise"
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    DirectoryListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/DirectoryMember"
        pagination:
          $ref: "#/components/schemas/Pagination"
    DirectoryMemberResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/DirectoryMember"

    # MEMBER SCHEMAS
    MembershipApplicationInput:
      type: object
      required: [fullName, age, cityId, stateId, institution, classYear]
      properties:
        fullName:
          type: string
        profilePhotoUrl:
          type: ["string", "null"]
        age:
          type: integer
          minimum: 13
          maximum: 35
        cityId:
          type: string
        stateId:
          type: string
        institution:
          type: string
        classYear:
          type: string
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        skillIds:
          type: array
          items:
            type: string
        privacySettings:
          $ref: "#/components/schemas/PrivacySettings"
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        profilePhotoUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        institution:
          type: string
        classYear:
          type: string
        skillIds:
          type: array
          items:
            type: string
        privacySettings:
          $ref: "#/components/schemas/PrivacySettings"
    PrivacySettings:
      type: object
      properties:
        phone:
          type: string
          enum: [public, private]
        whatsapp:
          type: string
          enum: [public, private]
        email:
          type: string
          enum: [public, private]
    MemberProfile:
      type: object
      required: [id, userId, status]
      properties:
        id:
          type: string
        userId:
          type: string
        memberId:
          type: ["string", "null"]
        fullName:
          type: string
        profilePhotoUrl:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        cityId:
          type: ["string", "null"]
        stateId:
          type: ["string", "null"]
        institution:
          type: ["string", "null"]
        classYear:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        rejectionReason:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        privacySettings:
          $ref: "#/components/schemas/PrivacySettings"
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        city:
          $ref: "#/components/schemas/City"
        state:
          $ref: "#/components/schemas/State"
        createdAt:
          type: string
...
    City:
      type: object
      required: [id, name, status]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [ACTIVE, UPCOMING]
        stateId:
          type: string
        state:
          $ref: "#/components/schemas/State"
        memberCount:
          type: integer
    State:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
...
    Skill:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
...
    ReportInput:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType:
          type: string
          enum: [POST, PROFILE]
        targetId:
          type: string
        reason:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]