# lib/api-zod/src/generated/types/directoryMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NetVent Parivaar API
 * OpenAPI spec version: 0.1.0
 */
import type { DirectoryMemberDegreeLevel } from './directoryMemberDegreeLevel';
import type { DirectoryMemberEducationType } from './directoryMemberEducationType';
import type { Expertise } from './expertise';

export interface DirectoryMember {
  id: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  educationType?: DirectoryMemberEducationType;
  /** @nullable */
  schoolOrCollegeName?: string | null;
  /** @nullable */
  schoolClass?: string | null;
  /** @nullable */
  degreeLevel?: DirectoryMemberDegreeLevel;
  /** @nullable */
  collegeYear?: string | null;
  isAvailable: boolean;
  expertise: Expertise[];
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
}